Exceptions move part1
Problem:
Currently exception classes are scattered all over SDK, this will lead to unecessary duplication and confusion

Analysis:
Removed UtilError exception class from mixin.py, corrected related files

Tests:
Flake8
