+++ f5/bigip/tm/util/Bash.py
@@ -27,8 +27,8 @@ REST Kind
 """
 
 from f5.bigip.mixins import CommandExecutionMixin
-from f5.bigip.mixins import UtilError
 from f5.bigip.resource import UnnamedResource
+from f5.sdk_exception import UtilError
 
 
 class Bash(UnnamedResource, CommandExecutionMixin):
