+++ f5/bigip/tm/util/test/functional/test_qkview.py
@@ -15,8 +15,8 @@
 
 import pytest
 
-from f5.bigip.mixins import UtilError
 from f5.bigip.resource import MissingRequiredCommandParameter
+from f5.sdk_exception import UtilError
 from icontrol.session import iControlUnexpectedHTTPError
