+++ f5/bigip/tm/util/test/functional/test_unix_mv.py
@@ -15,7 +15,7 @@
 
 import pytest
 
-from f5.bigip.mixins import UtilError
+from f5.sdk_exception import UtilError
 import os
 from tempfile import NamedTemporaryFile
